Added Image Uploader
The image uploader is pretty much done. Summary for the most important
features:

- User gets assigned a session which will expire once the user is
inactive for 30 mins
- User can access only images referenced with his sessionID
- With that logic, the user can still access his uploaded images when
he reloads the application or when he switches to a new algorithm
- The session and the image informations are stored in the mongoDB for
better image management on server site

Yet to be implemented:
- Hook into the session destroy prototype function from the MongoStore
(connect-mongo) for deleting the image folder assigned to that specific
sessionID. Otherwise we will get a lot of image garbage on the server.