added base64 encoded image to post request
now the post request sends a json object like follow:
{
  inputs: <Object>,
  highlighter: <Object>,
  image: <String>
}
whereas in the inputs field all the fields defined as input by that
algorithm (if any) are listed as key:value paris. the highlighter
contains information about the highlighted regions (if any) and the
image contains the image as base64 encoded string. the image format
will always be .png. make sure your server accepts requests with a
payload <= 10 mb.