Implemented #71
* some images are not valid png formatted and the creation of their
thumbnails fails. if this is the case, a warning is logged (only on
server side) and the path to the original result image is stored as
thumbUrl
* replaced lwip with gm (graphicsmagick) @lunactic @joapaspe please
make sure to install graphicsmagick on your machine and run npm install
(I suggest to also delete the node_modules folder before running npm
install).